pyproject.toml
[tool.poetry]
name = "dowload_photos"
version = "0.1.0"
description = ""
authors = ["Rasul Taimaskhanov <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
aiohttp = "^3.8.1"
loguru = "^0.6.0"
tortoise-orm = {extras = ["aiomysql"], version = "^0.19.0"}
aiofiles = "^0.8.0"
Pillow = "^9.1.0"
PyYAML = "^6.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
auto-py-to-exe = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
